All Error Solved (I think)
