+++ index.html
@@ -3,7 +3,7 @@
   <head>
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1">
-    <link rel="icon" type="image/svg+xml" href="/vite.svg" />
+    <link rel="icon" type="image/svg+xml" href="/src/assets/images/word_search icon.png" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <!-- <script>
       if(window.innerHeight > 630 && window.innerWidth > 830){
