+++ src/components/Home.jsx
@@ -9,11 +9,11 @@ function Home() {
     <div className="App1">
       <div className="header-container">
         <h1 className='head1'>Word Search</h1>
-        <img src="word_search icon.png" alt="Search Icon"  className='image'/>
+        <img src="../src/assets/images/word_search icon.png" alt="Search Icon"  className='image'/>
       </div>
       <h3>Welcome to the Puzzle Game</h3>
       <span>
-      <Link to="/word-matching/game1">
+      <Link to="/games/wordsearching/game1">
         <button className="Start-button">Start</button>
       </Link>
       <Link to="https://joywithlearning.com/games ">
