+++ src/components/PixiGame10.jsx
@@ -1,18 +1,18 @@
 import {Container, Stage, Text, Graphics } from '@pixi/react';
-import './App.css';
+import '../App.css';
 import '@pixi/events';
 import { useCallback, useEffect, useMemo, useState } from 'react';
 import { TextStyle } from 'pixi.js';
 import Confetti from 'react-confetti'
-import words from './words.json'
-// import ConfettiComponent from './ConfettiComponent';
-// import Confetti from 'canvas-confetti';
+import words from '../assets/words.json'
+import info_pic from '../assets/images/info_pic.png'
 
-function PixiGame6() {
+
+function PixiGame10() {
     
     const [dimensions, setDimensions] = useState({ width: window.innerWidth, height: window.innerHeight });
     const [puzzle, setPuzzle] = useState([]);
-    let completedWord = useMemo(()=>words[8].words, []);
+    let completedWord = useMemo(()=>words[4].words, []);
     const [selectedWord, setSelectedWord] = useState("")
     const [drawing, setDrawing] = useState(false)
     const [indices, setIndices] = useState([])
@@ -29,9 +29,10 @@ function PixiGame6() {
     const [mobile, setMobile] = useState(false)
     const [isCompleted, setIsCompleted] = useState(false)
     const [ongoingElapsedTime, setOngoingElapsedTime] = useState(0);
-    const [isStarted, setisStarted] = useState(false)
-    const textArr = useMemo(()=>words[8].grid,[]);
-    let hei
+    const textArr = useMemo(()=>words[4].grid,[]);
+    const [isStarted, setIsStarted] = useState(false)
+    let hei;
+
 
 
     const handleOrientationChange = useCallback(() => {
@@ -150,24 +151,30 @@ function PixiGame6() {
     
     useEffect(() => {
         if(window.innerHeight > 630 && window.innerWidth > 830){
-            let k = 0;
+        let k = 0;
         let newPuzzle = [];
         for (let ind = 0; ind < textArr.length; ind++) {
-            if (ind <= 10) {
+            if (ind <= 9) {
                 hei = 150;
-            } else if (ind <= 21) {
-                hei = 50;
-                if (ind === 11) {
+            } else if (ind <= 19) {
+                hei = 75;
+                if (ind === 10) {
                     k = 0;
                 }
-            } else {
-                hei = -50;
-                if (ind === 22) {
+            } else if(ind<=29){
+                hei = 0;
+                if (ind === 20) {
                     k = 0;
                 }
             }
+            else{
+                hei = -75;
+                if(ind ===30){
+                    k=0;
+                }
+            }
             let xPos = dimensions.width / 2 - 400 + 90 * (k);
-            let yPos = dimensions.height / 2 - hei;
+            let yPos = dimensions.height / 2 - hei - 80;
             k++;
 
             newPuzzle.push({
@@ -187,21 +194,27 @@ function PixiGame6() {
         let newPuzzle = [];
         setMobile(true)
         for (let ind = 0; ind < textArr.length; ind++) {
-            if (ind <= 10) {
-                hei = 50;
-            } else if (ind <= 21) {
-                hei = 0;
-                if (ind === 11) {
+            if (ind <= 9) {
+                hei = 150;
+            } else if (ind <= 19) {
+                hei = 75;
+                if (ind === 10) {
                     k = 0;
                 }
-            } else {
-                hei = -50;
-                if (ind === 22) {
+            } else if(ind<=29){
+                hei = 0;
+                if (ind === 20) {
                     k = 0;
                 }
             }
-            let xPos = dimensions.width / 2 - 300 + 50 * (k);
-            let yPos = dimensions.height / 2 - hei*2;
+            else{
+                hei = -75;
+                if(ind ===30){
+                    k=0;
+                }
+            }
+            let xPos = dimensions.width / 2 - 400 + 90 * (k);
+            let yPos = dimensions.height / 2 - hei - 80;
             k++;
 
             newPuzzle.push({
@@ -300,7 +313,7 @@ function PixiGame6() {
                     Math.pow(pointerPosition_y - letterPosition.y, 2)
                 );
                 if(distance< 40 && !mobile){
-                    setisStarted(true)
+                    setIsStarted(true)
                     setDrawing(true)
                         setSelectedWord(prev => prev + word.text)
                         setIndices([...indices, word.index])
@@ -312,7 +325,7 @@ function PixiGame6() {
                         };
             }
                 else if(distance <40 && mobile){
-                    setisStarted(true)
+                    setIsStarted(true);
                     setDrawing(true)
                         setSelectedWord(prev => prev + word.text)
                         setIndices([...indices, word.index])
@@ -445,14 +458,14 @@ if(window.innerHeight > 630 && window.innerWidth > 830){
             <div className="App">
     {/* {isCompleted && <ConfettiComponent isCompleted={{isCompleted}}/>} */}
       <div className="image-container">
-        <img src='../info_pic.png' alt="Descriptive Image" className="hover-image" onClick={()=>readOutLoud("Find the words listed below  Click and drag on the letters to select them")} style={{height: 35}}/>
+        <img src={info_pic} alt="Descriptive Image" className="hover-image" onClick={()=>readOutLoud("Find the words listed below  Click and drag on the letters to select them")} style={{height: 35}}/>
         <span style={{display: 'flex'}}><div className="description">Find the words listed below  Click and drag on the letters to select them.</div>
         {<h4 style={{marginLeft: window.innerWidth/2 + 80}}>Time: {ongoingElapsedTime.toFixed(0)}</h4>}</span>
       </div>
         {<div style={{marginLeft: window.innerWidth - 150}}><h4>Tries: {tries}</h4> </div>}
     </div>
 
-        <Stage x={0} y={0} options={{ backgroundColor: 11644879 }} height={dimensions.height - 250} width={dimensions.width}>
+        <Stage x={0} y={0} options={{ backgroundColor: 11644879}} height={dimensions.height - 250} width={dimensions.width}>
         <Graphics draw={draw} /> 
 
             <Container name='textArea'>
@@ -481,8 +494,8 @@ if(window.innerHeight > 630 && window.innerWidth > 830){
         <h3>Completed Words: {StrCompletedWords}</h3>
         <br/>
         <div style={{marginLeft: window.innerWidth/4 + 100}}>
-        <a type="button" className="btn btn-secondary btn-lg" href='/word-matching/game5'>Previous</a>
-            <a type="button" className="btn btn-secondary btn-lg" href='/word-matching/game7' style={{marginLeft: window.innerWidth/4}}>Next</a>
+        <a type="button" className="btn btn-secondary btn-lg" href='/games/wordsearching/game9'>Previous</a>
+            <a type="button" className="btn btn-secondary btn-lg" href='/games/wordsearching/Completed' style={{marginLeft: window.innerWidth/4}}>Next</a>
             </div>
             <br/>
             <br/>
@@ -501,9 +514,9 @@ else{
         {isCompleted && <Confetti/>}
         <div className="A">
       <div className="image-container">
-        <img src='../info_pic.png' alt="Descriptive Image" className="hover-image" onClick={()=>readOutLoud("Find the words listed below  Click and drag on the letters to select them")} style={{height: 35}}/>
+        <img src={info_pic} alt="Descriptive Image" className="hover-image" onClick={()=>readOutLoud("Find the words listed below  Click and drag on the letters to select them")} style={{height: 35}}/>
         <span style={{display: 'flex'}}><div className="description">Find the words listed below  Click and drag on the letters to select them.</div>
-        <h5 style={{marginLeft: window.innerWidth/15, padding: 0, width: 100}}>Time: {ongoingElapsedTime.toFixed(0)}</h5><h4 style={{marginLeft: window.innerWidth/12, padding: 0, width: 100}}>Tries: {tries}</h4></span>
+        <h4 style={{marginLeft: window.innerWidth/12}}>Tries: {tries}</h4></span>
       </div>
     </div>
         {/* <Voice ReadingText={"Find the words listed below  Click and drag on the letters to select them"}/> */}
@@ -535,7 +548,7 @@ else{
             <h1>Given Words: {givenWords}</h1>
             <h3>Completed Words: {StrCompletedWords}</h3>
             <div style={{marginLeft: window.innerWidth/4}}>
-        <a type="button" className="btn btn-secondary" href='/word-matching/game5'>Previous</a>
+        <a type="button" className="btn btn-secondary" href='/'>Previous</a>
             <a type="button" className="btn btn-secondary" href='/' style={{marginLeft: window.innerWidth/4}}>Next</a>
             </div>
 
@@ -548,4 +561,4 @@ else{
 }
 }
 
-export default PixiGame6;
+export default PixiGame10;
