+++ src/components/PixiGame6.jsx
@@ -1,18 +1,18 @@
 import {Container, Stage, Text, Graphics } from '@pixi/react';
-import './App.css';
+import '../App.css';
 import '@pixi/events';
 import { useCallback, useEffect, useMemo, useState } from 'react';
 import { TextStyle } from 'pixi.js';
 import Confetti from 'react-confetti'
-import words from './words.json'
-// import ConfettiComponent from './ConfettiComponent';
-// import Confetti from 'canvas-confetti';
+import words from '../assets/words.json'
+import info_pic from '../assets/images/info_pic.png'
 
-function PixiGame8() {
+
+function PixiGame6() {
     
     const [dimensions, setDimensions] = useState({ width: window.innerWidth, height: window.innerHeight });
     const [puzzle, setPuzzle] = useState([]);
-    let completedWord = useMemo(()=>words[9].words, []);
+    let completedWord = useMemo(()=>words[8].words, []);
     const [selectedWord, setSelectedWord] = useState("")
     const [drawing, setDrawing] = useState(false)
     const [indices, setIndices] = useState([])
@@ -30,9 +30,8 @@ function PixiGame8() {
     const [isCompleted, setIsCompleted] = useState(false)
     const [ongoingElapsedTime, setOngoingElapsedTime] = useState(0);
     const [isStarted, setisStarted] = useState(false)
-    const textArr = useMemo(()=>words[9].grid,[]);
-    let hei;
-
+    const textArr = useMemo(()=>words[8].grid,[]);
+    let hei
 
 
     const handleOrientationChange = useCallback(() => {
@@ -42,6 +41,13 @@ function PixiGame8() {
         }
     }, []);
 
+    useEffect(() => {
+        window.addEventListener("orientationchange", handleOrientationChange);
+        return () => {
+            window.removeEventListener("orientationchange", handleOrientationChange);
+        };
+    }, [handleOrientationChange]);
+
     const toggleScroll = (disable) => {
         if (disable) {
             document.body.style.overflow = 'hidden';
@@ -49,13 +55,6 @@ function PixiGame8() {
             document.body.style.overflow = 'auto';
         }
     };
-
-    useEffect(() => {
-        window.addEventListener("orientationchange", handleOrientationChange);
-        return () => {
-            window.removeEventListener("orientationchange", handleOrientationChange);
-        };
-    }, [handleOrientationChange]);
     
     
     function readOutLoud(text){
@@ -71,8 +70,7 @@ function PixiGame8() {
     }
 
     useEffect(() => {
-        if(isStarted)
-        setStartTime(new Date());
+        isStarted && setStartTime(new Date());
     }, [isStarted]);
 
     useEffect(() => {
@@ -152,30 +150,24 @@ function PixiGame8() {
     
     useEffect(() => {
         if(window.innerHeight > 630 && window.innerWidth > 830){
-        let k = 0;
+            let k = 0;
         let newPuzzle = [];
         for (let ind = 0; ind < textArr.length; ind++) {
-            if (ind <= 15) {
+            if (ind <= 10) {
                 hei = 150;
-            } else if (ind <= 31) {
-                hei = 75;
-                if (ind === 16) {
+            } else if (ind <= 21) {
+                hei = 50;
+                if (ind === 11) {
                     k = 0;
                 }
-            } else if(ind<=47){
-                hei = 0;
-                if (ind === 32) {
+            } else {
+                hei = -50;
+                if (ind === 22) {
                     k = 0;
                 }
             }
-            else{
-                hei = -75;
-                if(ind ===48){
-                    k=0;
-                }
-            }
-            let xPos = dimensions.width / 2 - 600 + 80 * (k);
-            let yPos = dimensions.height / 2 - hei - 60;
+            let xPos = dimensions.width / 2 - 400 + 90 * (k);
+            let yPos = dimensions.height / 2 - hei;
             k++;
 
             newPuzzle.push({
@@ -195,27 +187,21 @@ function PixiGame8() {
         let newPuzzle = [];
         setMobile(true)
         for (let ind = 0; ind < textArr.length; ind++) {
-            if (ind <= 15) {
-                hei = 150;
-            } else if (ind <= 31) {
-                hei = 75;
-                if (ind === 16) {
-                    k = 0;
-                }
-            } else if(ind<=47){
+            if (ind <= 10) {
+                hei = 50;
+            } else if (ind <= 21) {
                 hei = 0;
-                if (ind === 32) {
+                if (ind === 11) {
+                    k = 0;
+                }
+            } else {
+                hei = -50;
+                if (ind === 22) {
                     k = 0;
                 }
             }
-            else{
-                hei = -75;
-                if(ind ===48){
-                    k=0;
-                }
-            }
-            let xPos = dimensions.width / 2 - 400 + 50 * (k);
-            let yPos = dimensions.height / 2 - hei;
+            let xPos = dimensions.width / 2 - 300 + 50 * (k);
+            let yPos = dimensions.height / 2 - hei*2;
             k++;
 
             newPuzzle.push({
@@ -459,14 +445,14 @@ if(window.innerHeight > 630 && window.innerWidth > 830){
             <div className="App">
     {/* {isCompleted && <ConfettiComponent isCompleted={{isCompleted}}/>} */}
       <div className="image-container">
-        <img src='../info_pic.png' alt="Descriptive Image" className="hover-image" onClick={()=>readOutLoud("Find the words listed below  Click and drag on the letters to select them")} style={{height: 35}}/>
+        <img src={info_pic} alt="Descriptive Image" className="hover-image" onClick={()=>readOutLoud("Find the words listed below  Click and drag on the letters to select them")} style={{height: 35}}/>
         <span style={{display: 'flex'}}><div className="description">Find the words listed below  Click and drag on the letters to select them.</div>
         {<h4 style={{marginLeft: window.innerWidth/2 + 80}}>Time: {ongoingElapsedTime.toFixed(0)}</h4>}</span>
       </div>
         {<div style={{marginLeft: window.innerWidth - 150}}><h4>Tries: {tries}</h4> </div>}
     </div>
 
-        <Stage x={0} y={0} options={{ backgroundColor: 11644879}} height={dimensions.height - 250} width={dimensions.width}>
+        <Stage x={0} y={0} options={{ backgroundColor: 11644879 }} height={dimensions.height - 250} width={dimensions.width}>
         <Graphics draw={draw} /> 
 
             <Container name='textArea'>
@@ -495,8 +481,8 @@ if(window.innerHeight > 630 && window.innerWidth > 830){
         <h3>Completed Words: {StrCompletedWords}</h3>
         <br/>
         <div style={{marginLeft: window.innerWidth/4 + 100}}>
-        <a type="button" className="btn btn-secondary btn-lg" href='/word-matching/game7'>Previous</a>
-            <a type="button" className="btn btn-secondary btn-lg" href='/word-matching/game9' style={{marginLeft: window.innerWidth/4}}>Next</a>
+        <a type="button" className="btn btn-secondary btn-lg" href='/games/wordsearching/game5'>Previous</a>
+            <a type="button" className="btn btn-secondary btn-lg" href='/games/wordsearching/game7' style={{marginLeft: window.innerWidth/4}}>Next</a>
             </div>
             <br/>
             <br/>
@@ -515,9 +501,9 @@ else{
         {isCompleted && <Confetti/>}
         <div className="A">
       <div className="image-container">
-        <img src='../info_pic.png' alt="Descriptive Image" className="hover-image" onClick={()=>readOutLoud("Find the words listed below  Click and drag on the letters to select them")} style={{height: 35}}/>
+        <img src={info_pic} alt="Descriptive Image" className="hover-image" onClick={()=>readOutLoud("Find the words listed below  Click and drag on the letters to select them")} style={{height: 35}}/>
         <span style={{display: 'flex'}}><div className="description">Find the words listed below  Click and drag on the letters to select them.</div>
-        <h4 style={{marginLeft: window.innerWidth/12}}>Tries: {tries}</h4></span>
+        <h5 style={{marginLeft: window.innerWidth/15, padding: 0, width: 100}}>Time: {ongoingElapsedTime.toFixed(0)}</h5><h4 style={{marginLeft: window.innerWidth/12, padding: 0, width: 100}}>Tries: {tries}</h4></span>
       </div>
     </div>
         {/* <Voice ReadingText={"Find the words listed below  Click and drag on the letters to select them"}/> */}
@@ -549,7 +535,7 @@ else{
             <h1>Given Words: {givenWords}</h1>
             <h3>Completed Words: {StrCompletedWords}</h3>
             <div style={{marginLeft: window.innerWidth/4}}>
-        <a type="button" className="btn btn-secondary" href='/'>Previous</a>
+        <a type="button" className="btn btn-secondary" href='/games/wordsearching/game5'>Previous</a>
             <a type="button" className="btn btn-secondary" href='/' style={{marginLeft: window.innerWidth/4}}>Next</a>
             </div>
 
@@ -562,4 +548,4 @@ else{
 }
 }
 
-export default PixiGame8;
+export default PixiGame6;
