+++ src/components/PixiGame8.jsx
@@ -1,18 +1,18 @@
 import {Container, Stage, Text, Graphics } from '@pixi/react';
-import './App.css';
+import '../App.css';
 import '@pixi/events';
 import { useCallback, useEffect, useMemo, useState } from 'react';
 import { TextStyle } from 'pixi.js';
 import Confetti from 'react-confetti'
-import words from './words.json'
-// import ConfettiComponent from './ConfettiComponent';
-// import Confetti from 'canvas-confetti';
+import words from '../assets/words.json'
+import info_pic from '../assets/images/info_pic.png'
 
-function PixiGame5() {
+
+function PixiGame8() {
     
     const [dimensions, setDimensions] = useState({ width: window.innerWidth, height: window.innerHeight });
     const [puzzle, setPuzzle] = useState([]);
-    let completedWord = useMemo(()=>words[7].words, []);
+    let completedWord = useMemo(()=>words[9].words, []);
     const [selectedWord, setSelectedWord] = useState("")
     const [drawing, setDrawing] = useState(false)
     const [indices, setIndices] = useState([])
@@ -30,8 +30,9 @@ function PixiGame5() {
     const [isCompleted, setIsCompleted] = useState(false)
     const [ongoingElapsedTime, setOngoingElapsedTime] = useState(0);
     const [isStarted, setisStarted] = useState(false)
-    const textArr = useMemo(()=>words[7].grid,[]);
+    const textArr = useMemo(()=>words[9].grid,[]);
     let hei;
+
 
 
     const handleOrientationChange = useCallback(() => {
@@ -154,21 +155,27 @@ function PixiGame5() {
         let k = 0;
         let newPuzzle = [];
         for (let ind = 0; ind < textArr.length; ind++) {
-            if (ind <= 10) {
+            if (ind <= 15) {
                 hei = 150;
-            } else if (ind <= 21) {
-                hei = 50;
-                if (ind === 11) {
+            } else if (ind <= 31) {
+                hei = 75;
+                if (ind === 16) {
                     k = 0;
                 }
-            } else {
-                hei = -50;
-                if (ind === 22) {
+            } else if(ind<=47){
+                hei = 0;
+                if (ind === 32) {
                     k = 0;
                 }
             }
-            let xPos = dimensions.width / 2 - 400 + 90 * (k);
-            let yPos = dimensions.height / 2 - hei;
+            else{
+                hei = -75;
+                if(ind ===48){
+                    k=0;
+                }
+            }
+            let xPos = dimensions.width / 2 - 600 + 80 * (k);
+            let yPos = dimensions.height / 2 - hei - 60;
             k++;
 
             newPuzzle.push({
@@ -184,25 +191,31 @@ function PixiGame5() {
         setPuzzle(newPuzzle);
     }
     else{
-       let k = 0;
+        let k = 0;
         let newPuzzle = [];
         setMobile(true)
         for (let ind = 0; ind < textArr.length; ind++) {
-            if (ind <= 10) {
-                hei = 50;
-            } else if (ind <= 21) {
-                hei = 0;
-                if (ind === 11) {
+            if (ind <= 15) {
+                hei = 150;
+            } else if (ind <= 31) {
+                hei = 75;
+                if (ind === 16) {
                     k = 0;
                 }
-            } else {
-                hei = -50;
-                if (ind === 22) {
+            } else if(ind<=47){
+                hei = 0;
+                if (ind === 32) {
                     k = 0;
                 }
             }
-            let xPos = dimensions.width / 2 - 300 + 50 * (k);
-            let yPos = dimensions.height / 2 - hei*2;
+            else{
+                hei = -75;
+                if(ind ===48){
+                    k=0;
+                }
+            }
+            let xPos = dimensions.width / 2 - 400 + 50 * (k);
+            let yPos = dimensions.height / 2 - hei;
             k++;
 
             newPuzzle.push({
@@ -248,6 +261,7 @@ function PixiGame5() {
                         ...line,{start_x: puzzle[indices[indices.length - 2]].xPos, start_y: puzzle[indices[indices.length - 2]].yPos, end_x:puzzle[indices[indices.length - 1]].xPos, end_y:puzzle[indices[indices.length - 1]].yPos, clear: "none"}
                     ])
                 }
+                    // word.initColor = 'green'
                 }
                 return {
                     ...word,
@@ -445,14 +459,14 @@ if(window.innerHeight > 630 && window.innerWidth > 830){
             <div className="App">
     {/* {isCompleted && <ConfettiComponent isCompleted={{isCompleted}}/>} */}
       <div className="image-container">
-        <img src='../info_pic.png' alt="Descriptive Image" className="hover-image" onClick={()=>readOutLoud("Find the words listed below  Click and drag on the letters to select them")} style={{height: 35}}/>
+        <img src={info_pic} alt="Descriptive Image" className="hover-image" onClick={()=>readOutLoud("Find the words listed below  Click and drag on the letters to select them")} style={{height: 35}}/>
         <span style={{display: 'flex'}}><div className="description">Find the words listed below  Click and drag on the letters to select them.</div>
         {<h4 style={{marginLeft: window.innerWidth/2 + 80}}>Time: {ongoingElapsedTime.toFixed(0)}</h4>}</span>
       </div>
         {<div style={{marginLeft: window.innerWidth - 150}}><h4>Tries: {tries}</h4> </div>}
     </div>
 
-        <Stage x={0} y={0} options={{ backgroundColor: 11644879 }} height={dimensions.height - 250} width={dimensions.width}>
+        <Stage x={0} y={0} options={{ backgroundColor: 11644879}} height={dimensions.height - 250} width={dimensions.width}>
         <Graphics draw={draw} /> 
 
             <Container name='textArea'>
@@ -481,8 +495,8 @@ if(window.innerHeight > 630 && window.innerWidth > 830){
         <h3>Completed Words: {StrCompletedWords}</h3>
         <br/>
         <div style={{marginLeft: window.innerWidth/4 + 100}}>
-        <a type="button" className="btn btn-secondary btn-lg" href='/word-matching/game4'>Previous</a>
-            <a type="button" className="btn btn-secondary btn-lg" href='/word-matching/game6' style={{marginLeft: window.innerWidth/4}}>Next</a>
+        <a type="button" className="btn btn-secondary btn-lg" href='/games/wordsearching/game7'>Previous</a>
+            <a type="button" className="btn btn-secondary btn-lg" href='/games/wordsearching/game9' style={{marginLeft: window.innerWidth/4}}>Next</a>
             </div>
             <br/>
             <br/>
@@ -501,13 +515,13 @@ else{
         {isCompleted && <Confetti/>}
         <div className="A">
       <div className="image-container">
-        <img src='../info_pic.png' alt="Descriptive Image" className="hover-image" onClick={()=>readOutLoud("Find the words listed below  Click and drag on the letters to select them")} style={{height: 35}}/>
+        <img src={info_pic} alt="Descriptive Image" className="hover-image" onClick={()=>readOutLoud("Find the words listed below  Click and drag on the letters to select them")} style={{height: 35}}/>
         <span style={{display: 'flex'}}><div className="description">Find the words listed below  Click and drag on the letters to select them.</div>
-        <h5 style={{marginLeft: window.innerWidth/15, padding: 0, width: 100}}>Time: {ongoingElapsedTime.toFixed(0)}</h5><h4 style={{marginLeft: window.innerWidth/12, padding: 0, width: 100}}>Tries: {tries}</h4></span>
+        <h4 style={{marginLeft: window.innerWidth/12}}>Tries: {tries}</h4></span>
       </div>
     </div>
         {/* <Voice ReadingText={"Find the words listed below  Click and drag on the letters to select them"}/> */}
-            <Stage x={0} y={0} options={{ backgroundColor: 11644879 }} height={dimensions.height} width={dimensions.width - 150} className='stage-container'>
+            <Stage x={0} y={0} options={{ backgroundColor: 11644879}} height={dimensions.height} width={dimensions.width - 150} className='stage-container'>
             <Graphics draw={draw} /> 
     
                 <Container name='textArea'>
@@ -535,8 +549,8 @@ else{
             <h1>Given Words: {givenWords}</h1>
             <h3>Completed Words: {StrCompletedWords}</h3>
             <div style={{marginLeft: window.innerWidth/4}}>
-        <a type="button" className="btn btn-secondary" href='/word-matching/game3'>Previous</a>
-            <a type="button" className="btn btn-secondary" href='/word-matching/game6' style={{marginLeft: window.innerWidth/4}}>Next</a>
+        <a type="button" className="btn btn-secondary" href='/'>Previous</a>
+            <a type="button" className="btn btn-secondary" href='/' style={{marginLeft: window.innerWidth/4}}>Next</a>
             </div>
 
             {/* <h3>Time: {elapsedSeconds}</h3>
@@ -548,4 +562,4 @@ else{
 }
 }
 
-export default PixiGame5;
+export default PixiGame8;
